Passing query params from search to landing page
