+++ web/src/components/DandisetsPage.vue
@@ -100,27 +100,9 @@ import DandisetList from '@/components/DandisetList.vue';
 import DandisetSearchField from '@/components/DandisetSearchField.vue';
 import { dandiRest } from '@/rest';
 import { Dandiset, Paginated } from '@/types';
+import { sortingOptions } from '@/utils/constants';
 
 const DANDISETS_PER_PAGE = 8;
-
-const sortingOptions = [
-  {
-    name: 'Modified',
-    djangoField: 'modified',
-  },
-  {
-    name: 'Identifier',
-    djangoField: 'id',
-  },
-  {
-    name: 'Name',
-    djangoField: 'name',
-  },
-  {
-    name: 'Size',
-    djangoField: 'size',
-  },
-];
 
 export default defineComponent({
   name: 'DandisetsPage',
@@ -159,6 +141,8 @@ export default defineComponent({
 
     const djangoDandisetRequest: Ref<Paginated<Dandiset> | null> = ref(null);
     watchEffect(async () => {
+      // console.log('page', ctx.root.$route);
+
       const ordering = ((sortDir.value === -1) ? '-' : '') + sortField.value;
       const response = await dandiRest.dandisets({
         page: page.value,
